+++ hawtdb-engine/src/main/java/org/fusesource/hawtdb/internal/page/transaction/TxPageFileFactory.java
@@ -110,7 +110,7 @@ public class TxPageFileFactory {
         return pageFileFactory.getMaxPages();
     }
 
-    public short getPageSize() {
+    public int getPageSize() {
         return pageFileFactory.getPageSize();
     }
